Support specified shape key animations
A ShapeKeys node is used to specify "blend shapes" with weight keys, eg:

    ShapeKeys {
	noot = 0.5
	noot = 1
    }

will set up a single blend shape with 2 keys (really 3 as the basis
shape is implicitly available (but a key with weight 0 should override
it)). Using a different name for the value produces a different set of
keys. Note that keys with the same name must all be grouped together in
ascending weight order (Unity's rules). And the number of keys in the
node needs to match the number of delta shapes (not including the basis)
in the .mu file.